@@ -13,18 +13,20 @@ import os
 import re
 import sys
 
-sys.path.insert(0, os.path.abspath('..'))
-sys.path.append(os.path.abspath('extensions'))
+sys.path.insert(0, os.path.abspath(".."))
+sys.path.append(os.path.abspath("extensions"))
 
 
 # -- Project information -----------------------------------------------------
 
-project = 'steam community market'
-copyright = '2020, offish'
-author = 'offish'
+project = "steam community market"
+copyright = "2020, offish"
+author = "offish"
 
-with open('../steam_community_market/__init__.py') as f:
-    version = re.search(r'^__version__\s*=\s*[\'"]([^\'"]*)[\'"]', f.read(), re.MULTILINE).group(1)
+with open("../steam_community_market/__init__.py") as f:
+    version = re.search(
+        r'^__version__\s*=\s*[\'"]([^\'"]*)[\'"]', f.read(), re.MULTILINE
+    )[1]
 
 # The full version, including alpha/beta/rc tags
 release = version
@@ -35,55 +37,52 @@ release = version
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-    'builder',
-    'sphinx.ext.autodoc',
-    'sphinx.ext.extlinks',
-    'sphinx.ext.intersphinx',
-    'sphinx.ext.napoleon',
-    'details',
-    'exception_hierarchy'
+    "builder",
+    "sphinx.ext.autodoc",
+    "sphinx.ext.extlinks",
+    "sphinx.ext.intersphinx",
+    "sphinx.ext.napoleon",
+    "details",
+    "exception_hierarchy",
 ]
 
-sphinx_mapping = {
-    'py': ('https://docs.python.org/3', None)
-}
+sphinx_mapping = {"py": ("https://docs.python.org/3", None)}
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
+templates_path = ["_templates"]
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 
-master_doc = 'index'
+master_doc = "index"
 
 # -- Options for HTML output -------------------------------------------------
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'alabaster'
+html_theme = "alabaster"
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+html_static_path = ["_static"]
 
 
-autodoc_member_order = 'bysource'
+autodoc_member_order = "bysource"
 
 
 html_theme_options = {
-    'description': 'by offish',
-    'github_user': 'offish',
-    'github_repo': 'steam_community_market',
-    'github_button': 'true',
-    'github_banner': 'true',
-    'fixed_sidebar': 'true',
-    'donate_url': 'https://steamcommunity.com/tradeoffer/new/?partner=293059984&token=0-l_idZR',
+    "description": "by offish",
+    "github_user": "offish",
+    "github_repo": "steam_community_market",
+    "github_button": "true",
+    "github_banner": "true",
+    "fixed_sidebar": "true",
+    "donate_url": "https://steamcommunity.com/tradeoffer/new/?partner=293059984&token=0-l_idZR",
 }
-
 
 
 # Skal endres senere.
