@@ -7,43 +7,46 @@ class DPYStandaloneHTMLBuilder(StandaloneHTMLBuilder):
     def write_genindex(self) -> None:
         # the total count of lines for each index letter, used to distribute
         # the entries into two columns
-        genindex = IndexEntries(self.env).create_index(
-            self, group_entries=False)
-        indexcounts = []
-        for _k, entries in genindex:
-            indexcounts.append(sum(1 + len(subitems)
-                                   for _, (_, subitems, _) in entries))
+        genindex = IndexEntries(self.env).create_index(self, group_entries=False)
+        indexcounts = [
+            sum(1 + len(subitems) for _, (_, subitems, _) in entries)
+            for _, entries in genindex
+        ]
 
         genindexcontext = {
-            'genindexentries': genindex,
-            'genindexcounts': indexcounts,
-            'split_index': self.config.html_split_index,
+            "genindexentries": genindex,
+            "genindexcounts": indexcounts,
+            "split_index": self.config.html_split_index,
         }
 
         if self.config.html_split_index:
-            self.handle_page('genindex', genindexcontext,
-                             'genindex-split.html')
-            self.handle_page('genindex-all', genindexcontext,
-                             'genindex.html')
+            self.handle_page("genindex", genindexcontext, "genindex-split.html")
+            self.handle_page("genindex-all", genindexcontext, "genindex.html")
+
             for (key, entries), count in zip(genindex, indexcounts):
-                ctx = {'key': key, 'entries': entries, 'count': count,
-                       'genindexentries': genindex}
-                self.handle_page('genindex-' + key, ctx,
-                                 'genindex-single.html')
+                ctx = {
+                    "key": key,
+                    "entries": entries,
+                    "count": count,
+                    "genindexentries": genindex,
+                }
+                self.handle_page(f"genindex-{key}", ctx, "genindex-single.html")
         else:
-            self.handle_page('genindex', genindexcontext, 'genindex.html')
+            self.handle_page("genindex", genindexcontext, "genindex.html")
 
 
 def get_builder(app):
     """This is necessary because RTD injects their own for some reason."""
     try:
-        original = app.registry.builders['readthedocs']
+        original = app.registry.builders["readthedocs"]
     except KeyError:
         return DPYStandaloneHTMLBuilder
     else:
-        injected_mro = tuple(base if base is not StandaloneHTMLBuilder else DPYStandaloneHTMLBuilder
-                             for base in original.mro()[1:])
-        return type(original.__name__, injected_mro, {'name': 'readthedocs'})
+        injected_mro = tuple(
+            base if base is not StandaloneHTMLBuilder else DPYStandaloneHTMLBuilder
+            for base in original.mro()[1:]
+        )
+        return type(original.__name__, injected_mro, {"name": "readthedocs"})
 
 
 def setup(app):
